♻️ [Refactor] 리액트쿼리 응집화
