+++ src/main/ipcHandler.ts
@@ -6,35 +6,35 @@ import { store } from './store';
 import activeWindow from 'active-win';
 
 export const registerIPCHandlers = () => {
-    ipcMain.handle('try-auto-login', tryAutoLogin);
-    ipcMain.handle('logout-google-oauth', logoutGoogleOAuth);
-    ipcMain.on('start-google-oauth', startGoogleOAuth);
+  ipcMain.handle('try-auto-login', tryAutoLogin);
+  ipcMain.handle('logout-google-oauth', logoutGoogleOAuth);
+  ipcMain.on('start-google-oauth', startGoogleOAuth);
 
-    ipcMain.on('quit-app', () => {
-        app.quit();
-    });
+  ipcMain.on('quit-app', () => {
+    app.quit();
+  });
 
-    ipcMain.on('start-dragging', () => {
-        detach(mainWindow);
-        mainWindow.setResizable(true);
-    });
-    ipcMain.on('stop-dragging', () => {
-        mainWindow.setResizable(false);
-        const { width, height, x, y } = mainWindow.getBounds();
-        attach(mainWindow, { forwardKeyboardInput: true, forwardMouseInput: true });
-        store.set('window-bounds', { width, height, x, y });
-    });
+  ipcMain.on('start-dragging', () => {
+    detach(mainWindow);
+    mainWindow.setResizable(true);
+  });
+  ipcMain.on('stop-dragging', () => {
+    mainWindow.setResizable(false);
+    const { width, height, x, y } = mainWindow.getBounds();
+    attach(mainWindow, { forwardKeyboardInput: true, forwardMouseInput: true });
+    store.set('window-bounds', { width, height, x, y });
+  });
 
-    ipcMain.on('set-opacity', (_, newOpacity) => {
-        mainWindow.setOpacity(newOpacity);
-        store.set('window-opacity', newOpacity);
-    });
+  ipcMain.on('set-opacity', (_, newOpacity) => {
+    mainWindow.setOpacity(newOpacity);
+    store.set('window-opacity', newOpacity);
+  });
 
-    ipcMain.handle('get-initial-opacity', () => store.get('window-opacity'));
+  ipcMain.handle('get-initial-opacity', () => store.get('window-opacity'));
 
-    ipcMain.on('renderer-ready', async (event) => {
-        const window = await activeWindow();
-        const isExplorer = window?.title === 'Program Manager';
-        event.sender.send('update-clickable', isExplorer);
-    });
+  ipcMain.on('renderer-ready', async (event) => {
+    const window = await activeWindow();
+    const isExplorer = window?.title === 'Program Manager';
+    event.sender.send('update-clickable', isExplorer);
+  });
 };
