+++ src/renderer/entities/event/hooks/useAddEvent.tsx
@@ -2,6 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { CalendarEvent } from '@/shared/types/EventType';
 import { createEventBody } from '@/features/event/lib/createEventBody';
 import { eventApi } from '@/entities/event';
+import { eventKeys } from '../api/queries';
 
 export function useAddEvent() {
   const queryClient = useQueryClient();
@@ -9,27 +10,23 @@ export function useAddEvent() {
   const addEventMutation = useMutation({
     mutationFn: eventApi.create,
     onMutate: async (newEvent) => {
-      await queryClient.cancelQueries({ queryKey: ['googleCalendarEvents'] });
-      const previousData = queryClient.getQueryData<{ items: CalendarEvent[] }>(['googleCalendarEvents']);
+      await queryClient.cancelQueries({ queryKey: eventKeys.events });
+      const previousData = queryClient.getQueryData<{ items: CalendarEvent[] }>(eventKeys.events);
       const newEventItem = {
         ...createEventBody(newEvent.eventData),
         id: `temp-id-${Date.now()}`
       };
 
       if (previousData) {
-        queryClient.setQueryData(['googleCalendarEvents'], {
+        queryClient.setQueryData(eventKeys.events, {
           items: [...previousData.items, newEventItem]
         });
       }
 
       return { previousData };
     },
-    onError: (_error, _variables, context) => {
-      if (context?.previousData) {
-        queryClient.setQueryData(['googleCalendarEvents'], context.previousData);
-      }
-    },
-    onSettled: () => queryClient.invalidateQueries({ queryKey: ['googleCalendarEvents'] })
+    onError: (_error, _variables, context) => queryClient.setQueryData(eventKeys.events, context?.previousData),
+    onSettled: () => queryClient.invalidateQueries({ queryKey: eventKeys.events })
   });
 
   return { addEvent: addEventMutation.mutate };
