+++ src/renderer/entities/event/hooks/useCompleteEvent.tsx
@@ -2,6 +2,7 @@ import { useLogin } from '@/shared/hooks/useLogin';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { CalendarEvent } from '@/shared/types/EventType';
 import { eventApi } from '@/entities/event';
+import { eventKeys } from '../api/queries';
 
 interface CompleteEventProp {
   eventId: string;
@@ -12,7 +13,6 @@ export function useCompleteEvent() {
   const queryClient = useQueryClient();
 
   const completeEventMutation = useMutation({
-    mutationKey: ['completeEvent'],
     mutationFn: async ({ eventId, isCompleted }: CompleteEventProp) => {
       const patchBody = {
         extendedProperties: {
@@ -24,11 +24,11 @@ export function useCompleteEvent() {
       return eventApi.complete({ accessToken: tokens.access_token, eventId, patchBody });
     },
     onMutate: async ({ eventId, isCompleted }) => {
-      await queryClient.cancelQueries({ queryKey: ['googleCalendarEvents'] });
-      const previousData = queryClient.getQueryData<{ items: CalendarEvent[] }>(['googleCalendarEvents']);
+      await queryClient.cancelQueries({ queryKey: eventKeys.events });
+      const previousData = queryClient.getQueryData<{ items: CalendarEvent[] }>(eventKeys.events);
 
       if (previousData) {
-        queryClient.setQueryData(['googleCalendarEvents'], {
+        queryClient.setQueryData(eventKeys.events, {
           items: previousData.items.map((item) => {
             if (item.id === eventId) {
               return {
@@ -47,13 +47,9 @@ export function useCompleteEvent() {
 
       return { previousData };
     },
-    onError: (_err, _variables, context) => {
-      if (context?.previousData) {
-        queryClient.setQueryData(['googleCalendarEvents'], context.previousData);
-      }
-    },
+    onError: (_err, _variables, context) => queryClient.setQueryData(eventKeys.events, context?.previousData),
     onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['googleCalendarEvents'] });
+      queryClient.invalidateQueries({ queryKey: eventKeys.events });
     }
   });
