+++ src/renderer/features/event/index.tsx
@@ -1,7 +1,7 @@
-export { HolidayButton } from './ui/holiday/HolidayButton';
+export { HolidayButton } from './ui/HolidayButton';
 
-export { AddEventForm } from './ui/add-event/AddEventForm';
-export { DeleteEventButton } from './ui/delete-event/DeleteEventButton';
-export { EditEventForm } from './ui/edit-event/EditEventForm';
-export { PalletteDropdown } from './ui/pallette/PalletteDropdown';
-export { CompleteEventButton } from './ui/complete-event/CompleteEventButton';
+export { AddEventForm } from './ui/AddEventForm';
+export { DeleteEventButton } from './ui/DeleteEventButton';
+export { EditEventForm } from './ui/EditEventForm';
+export { PalletteDropdown } from './ui/PalletteDropdown';
+export { CompleteEventButton } from './ui/CompleteEventButton';
