+++ src/renderer/features/event/types/FormType.ts
@@ -1,7 +1,7 @@
 export interface FormState {
-    summary: string;
-    colorId: string;
-    start: string;
-    end: string;
-    allDay: boolean;
+  summary: string;
+  colorId: string;
+  start: string;
+  end: string;
+  allDay: boolean;
 }
