+++ src/renderer/features/event/ui/DeleteEventButton.tsx
